requirements.txt
asgiref==3.2.3
beautifulsoup4==4.9.1
boto3==1.16.44
botocore==1.19.44
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap4==2.0.1
django-cors-headers==3.6.0
django-crispy-forms==1.8.1
django-filter==2.4.0
django-heroku==0.3.1
django-pwa==1.0.10
django-session-timeout==0.1.0
django-storages==1.11.1
django-use-email-as-username==1.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
jmespath==0.10.0
Markdown==3.3.3
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.7.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.0
stripe==2.55.1
urllib3==1.25.10
whitenoise==5.2.0
django-tinymce==3.3.0
django-allauth==0.45.0
django-filter-21.1
django-recaptcha-2.0.6
stripe==2.63.0